+++ IOAudioFamily/IOAudioTimeIntervalFilter.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998-2012 Apple Computer, Inc. All rights reserved.
+ * Copyright (c) 1998-2014 Apple Computer, Inc. All rights reserved.
  *
  * @APPLE_LICENSE_HEADER_START@
  *
@@ -208,7 +208,7 @@ public:
      */
     virtual void free();
 	
-	IOReturn reInitialiseFilter(uint64_t expectedInterval = 0, uint32_t multiIntervalCount = 1 );
+	IOReturn reInitialiseFilter(uint32_t expectedInterval = 0, uint32_t multiIntervalCount = 1 );
 	
 	OSMetaClassDeclareReservedUnused (IOAudioTimeIntervalFilterFIR, 0 );
 	OSMetaClassDeclareReservedUnused (IOAudioTimeIntervalFilterFIR, 1 );
