+++ ksh/ksh/src/lib/libast/include/stk.h
@@ -3,12 +3,12 @@
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
 *                      and is licensed under the                       *
-*                  Common Public License, Version 1.0                  *
+*                 Eclipse Public License, Version 1.0                  *
 *                    by AT&T Intellectual Property                     *
 *                                                                      *
 *                A copy of the License is available at                 *
-*            http://www.opensource.org/licenses/cpl1.0.txt             *
-*         (with md5 checksum 059e8cd6165cb4c31e351f2b69388fd9)         *
+*          http://www.eclipse.org/org/documents/epl-v10.html           *
+*         (with md5 checksum b35adb5213ca9657e911e9befb180842)         *
 *                                                                      *
 *              Information and Software Systems Research               *
 *                            AT&T Research                             *
@@ -66,12 +66,12 @@ extern Stk_t*		stkopen(int);
 extern Stk_t*		stkinstall(Stk_t*, char*(*)(int));
 extern int		stkclose(Stk_t*);
 extern int		stklink(Stk_t*);
-extern char*		stkalloc(Stk_t*, unsigned);
-extern char*		stkcopy(Stk_t*,const char*);
-extern char*		stkset(Stk_t*, char*, unsigned);
-extern char*		_stkseek(Stk_t*, unsigned);
-extern char*		stkfreeze(Stk_t*, unsigned);
-extern int		stkon(Stk_t*, char *);
+extern char*		stkalloc(Stk_t*, size_t);
+extern char*		stkcopy(Stk_t*, const char*);
+extern char*		stkset(Stk_t*, char*, size_t);
+extern char*		_stkseek(Stk_t*, ssize_t);
+extern char*		stkfreeze(Stk_t*, size_t);
+extern int		stkon(Stk_t*, char*);
 
 #undef	extern
